+++ kernel/Makefile
@@ -21,10 +21,12 @@ LIBS:=$(LIBS) -nostdlib -lk -lgcc
 ARCHDIR=arch/$(HOSTARCH)
 MMUDIR=mmu
 DEVICEDIR=arch/devices
+KERNELDIR=kernel
 
 include $(ARCHDIR)/make.config
 include ${MMUDIR}/make.config
 include ${DEVICEDIR}/make.config
+include ${KERNELDIR}/make.config
 
 CFLAGS:=$(CFLAGS) $(KERNEL_ARCH_CFLAGS)
 CPPFLAGS:=$(CPPFLAGS) $(KERNEL_ARCH_CPPFLAGS)
@@ -33,7 +35,7 @@ LIBS:=$(LIBS) $(KERNEL_ARCH_LIBS)
 
 KERNEL_OBJS=\
 $(KERNEL_ARCH_OBJS) \
-kernel/kernel.o \
+${KERNEL_KERNEL_OBJS} \
 ${MMU_OBJS} \
 ${KERNEL_DEVICE_OBJS}
 
@@ -45,6 +47,7 @@ $(ARCHDIR)/crtend.o \
 $(ARCHDIR)/crtn.o \
 ${MMU_OBJS} \
 ${KERNEL_DEVICE_OBJS} \
+${KERNEL_KERNEL_OBJS} \
 
 LINK_LIST=\
 $(LDFLAGS) \
