+++ kernel/arch/i386/x86.c
@@ -16,18 +16,6 @@ void load_idt(struct idtr *idtr) {
        asm volatile("lidt %0" : : "m" (*idtr));
 }
 
-void outb(uint32_t ad, uint8_t v)
-{
-	asm volatile("outb %%al, %%dx" :: "d" (ad), "a" (v));
-}
-
-uint8_t inb(uint32_t ad)
-{
-	uint8_t v; 
-	asm volatile("inb %%dx, %%al" : "=a" (v) : "d" (ad));
-	return v;
-}
-
 void gdt_init(void) {
 
 	create_gdt_descriptor(0x0, 0x0, 0x0, 0x0, &kgdt[0]);
