+++ kernel/kernel/make.config
@@ -1,0 +1,9 @@
+KERNEL_ARCH_CFLAGS=
+KERNEL_ARCH_CPPFLAGS=
+KERNEL_ARCH_LDFLAGS=
+KERNEL_ARCH_LIBS=
+
+KERNEL_KERNEL_OBJS=\
+$(KERNELDIR)/kernel.o \
+$(KERNELDIR)/io.o \
+
